+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.8.2] - 2023-03-16
+
+### Added
+
+* Counter `EntryPoints_Tas_UncountedQueries`, to count how many queries we were not able to parse table information from.
+
+* Table access statistics can parse tables on queries containing `DATABASE()` function.
+
 ## [2.8.1] - 2023-03-16
 
 ### Fixed